@@ -1,12 +1,13 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 export class MessagesAlertContainer extends React.Component {
 
   static propTypes = {
-    mailCount: React.PropTypes.number.isRequired,
-    settingsOnClick: React.PropTypes.func,
-    allMessagesOnClick: React.PropTypes.func,
-    markAllAsReadOnClick: React.PropTypes.func,
+    mailCount: PropTypes.number.isRequired,
+    settingsOnClick: PropTypes.func,
+    allMessagesOnClick: PropTypes.func,
+    markAllAsReadOnClick: PropTypes.func,
   }
 
   constructor(props) {
@@ -37,13 +38,13 @@ export class MessagesAlertContainer extends React.Component {
             <i className='dropdown-arr'/>
             <div className='header clearfix'>
               <strong className='red-text'>Messages</strong>
-                {this.props.markAllAsReadOnClick ? <a onClick={this.props.markAllAsReadOnClick}>Mark All as Read</a> : ''}
-                {this.props.settingsOnClick ? <a onClick={this.props.settingsOnClick}>Settings</a> : ''}
+              {this.props.markAllAsReadOnClick ? <a onClick={this.props.markAllAsReadOnClick}>Mark All as Read</a> : ''}
+              {this.props.settingsOnClick ? <a onClick={this.props.settingsOnClick}>Settings</a> : ''}
             </div>
             <div className='msg-list'>
               {this.props.children}
             </div>
-              {this.props.allMessagesOnClick ? <a onClick={this.props.allMessagesOnClick}>See all messages</a> : ''}
+            {this.props.allMessagesOnClick ? <a onClick={this.props.allMessagesOnClick}>See all messages</a> : ''}
           </div>
         </li>
       </ul>
