@@ -1,12 +1,13 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 export class Panel extends React.Component {
 
   static propTypes = {
-    title: React.PropTypes.string,
-    className: React.PropTypes.string,
-    size: React.PropTypes.string,
-    withScroll: React.PropTypes.bool,
+    title: PropTypes.string,
+    className: PropTypes.string,
+    size: PropTypes.string,
+    withScroll: PropTypes.bool,
   }
 
   static defaultProps = {
@@ -16,7 +17,7 @@ export class Panel extends React.Component {
   }
 
   renderHeader() {
-    if (! this.props.title) {
+    if (!this.props.title) {
       return null;
     }
 
@@ -31,32 +32,33 @@ export class Panel extends React.Component {
 
   renderPanelSize() {
     switch (this.props.size) {
-    case 'xs':
-    case 'extra-small':
-      return 'xsmall-panel';
+      case 'xs':
+      case 'extra-small':
+        return 'xsmall-panel';
 
-    case 'sm':
-    case 'small':
-      return 'small-panel';
+      case 'sm':
+      case 'small':
+        return 'small-panel';
 
-    case 'md':
-    case 'medium':
-      return 'medium-panel';
+      case 'md':
+      case 'medium':
+        return 'medium-panel';
 
-    case 'lg':
-    case 'large':
-      return 'large-panel';
+      case 'lg':
+      case 'large':
+        return 'large-panel';
 
-    case 'auto':
-    case 'none':
-    default:
-      return '';
+      case 'auto':
+      case 'none':
+      default:
+        return '';
     }
   }
 
   render() {
     return (
-      <div className={`panel panel-blur ${this.renderPanelSize()} light-text ${this.props.withScroll ? 'with-scroll' : ''} ${this.props.className}`}>
+      <div
+        className={`panel panel-blur ${this.renderPanelSize()} light-text ${this.props.withScroll ? 'with-scroll' : ''} ${this.props.className}`}>
         {this.renderHeader()}
         <div className='panel-body'>
           {this.props.children}
