@@ -1,13 +1,14 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
-export const Tab = ({ title, onClick }) => {
+export const Tab = ({title, onClick}) => {
   return (
     <a href='#' onClick={onClick}>{title}</a>
   );
 };
 
 Tab.propTypes = {
-  title: React.PropTypes.string,
-  onClick: React.PropTypes.func,
-  name: React.PropTypes.string,
+  title: PropTypes.string,
+  onClick: PropTypes.func,
+  name: PropTypes.string,
 };
