@@ -1,9 +1,10 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 export class Loading extends React.Component {
   render() {
     return (
-      <i className='fa fa-spinner fa-spin' />
+      <i className='fa fa-spinner fa-spin'/>
     );
   }
 }
