@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import React from 'react';
+import React from 'react'; import PropTypes from 'prop-types';
 import {shallow} from 'enzyme';
 import {expect} from 'chai';
 
