@@ -1,4 +1,4 @@
-import React from 'react';
+import React from 'react'; import PropTypes from 'prop-types';
 import {expect} from 'chai';
 import {mount} from 'enzyme';
 import { noop } from 'lodash';
