@@ -1,4 +1,4 @@
-import React from 'react';
+import React from 'react'; import PropTypes from 'prop-types';
 import sinon from 'sinon';
 import {noop} from 'lodash';
 import {expect} from 'chai';
